moved common to utils+
