--- cfg/vio.cfg
@@ -1,49 +1,0 @@
-######### General #########
-max_iter 10
-th_iter 0.1
-use_sim 0
-timeOffsetCam 0
-timeOffsetGyr -0.05
-timeOffsetAcc 0
-do_draw 1
-draw_tracking 1
-draw_adding 0
-single_match_only 0
-
-######### Residual weighting #########
-w_posfd		1e2
-w_attfd		1e3
-w_accpre	1e1
-w_gyrupd	1e3
-w_accbias	1e3
-w_gyrbias	1e3
-w_imgupd	1e2
-w_beapre	1e2
-w_dispre	1e2
-w_veppre	1e4
-w_veapre	1e4
-
-######### State initialization #########
-init_vep 		-0.0111674199187 -0.0574640920022 0.0207586947896
-init_vea 		0.712115587266 0.00666398307551 -0.0079168224269 -0.701985972528
-init_dis		0.5
-initstd_pos		1e2
-initstd_att		1e1
-initstd_vel		1e1
-initstd_ror		1e1
-initstd_acb		1e1
-initstd_gyb		1e1
-initstd_vep		1e3
-initstd_vea		1e3
-initstd_bea		1e2
-initstd_dis		1e0
-
-######### Landmark tracking #########
-match_desc_distance 	100
-match_geom_distance 	3
-prune_count 			5
-num_candidates_matching	10
-num_candidates_add		500
-bucket_count			6
-neighbor_suppression	50
-start_add				5
