+++ include/tsif/element_vector.h
@@ -1,7 +1,7 @@
 #ifndef TSIF_ELEMENT_VECTOR_H_
 #define TSIF_ELEMENT_VECTOR_H_
 
-#include "tsif/common.h"
+#include "tsif/utils/common.h"
 #include "tsif/element.h"
 
 namespace tsif{
