+++ include/tsif/filters/imu_gps.h
@@ -7,7 +7,7 @@
 #include "tsif/residuals/pose_update.h"
 #include "tsif/residuals/position_findif.h"
 #include "tsif/residuals/random_walk.h"
-#include "tsif/common.h"
+#include "tsif/utils/common.h"
 #include "tsif/filter.h"
 
 namespace tsif{
