+++ include/tsif/residuals/gyroscope_update.h
@@ -1,7 +1,7 @@
 #ifndef TSIF_GYROSCOPE_UPDATE_H_
 #define TSIF_GYROSCOPE_UPDATE_H_
 
-#include "tsif/common.h"
+#include "tsif/utils/common.h"
 #include "tsif/residual.h"
 
 namespace tsif{
