+++ include/tsif/residuals/random_walk.h
@@ -1,7 +1,7 @@
 #ifndef TSIF_RANDOM_WALK_H_
 #define TSIF_RANDOM_WALK_H_
 
-#include "tsif/common.h"
+#include "tsif/utils/common.h"
 #include "tsif/residual.h"
 
 namespace tsif{
