+++ src/test_quaternion.cpp
@@ -1,4 +1,4 @@
-#include "tsif/common.h"
+#include "tsif/utils/common.h"
 #include "tsif/element.h"
 
 using namespace tsif;
